[Chapter01:Section6] state 만드는 부분 완전히 별도 함수로 분리
